Refactor with `Types.py`
